refactor(capsule-install): extract install machinery into its own crate
The CLI install path used to be 2371 lines in
`astrid-cli/src/commands/capsule/install.rs` mixing source resolution
(GitHub release-asset download, clone+build, OpenClaw transpile,
archive auto-detect) with the actual install machinery (file layout,
content-addressing of WASM/WIT, lifecycle hooks, topic baking,
meta.json writes). The kernel-side InstallCapsule handler couldn't
share that code because astrid-kernel can't depend on astrid-cli.

This commit pulls the post-resolution install machinery into a new
`astrid-capsule-install` crate that both the CLI and the kernel will
consume. CLI source resolution stays where it was. Behaviour is
preserved end-to-end (all CLI install paths still route through to
the lib at the same point they used to).

While extracting, two improvements:

* Source-direct content-addressing. Previously the install copied
  the entire capsule tree into target_dir, then re-read the .wasm
  to hash it, wrote bin/<hash>.wasm, and deleted the per-capsule
  copy — same dance for wit/. The lib now hashes WASM and WIT from
  the source directly, writes once with atomic temp+rename to the
  content store, and the per-capsule directory copy excludes *.wasm
  and the top-level wit/ by construction. The runtime contract is
  unchanged (loader still resolves by hash); only the install path
  is cleaner. Pre-flight hashing also means a content-addressing
  failure leaves the existing install intact.
* Lifecycle hook event-bus boundary. run_lifecycle now accepts an
  optional external EventBus instead of always creating its own.
  The CLI passes one with a stdin elicit handler subscribed
  (preserving install-time elicit prompts). The kernel-side handler
  will pass None — no human at the daemon end to answer prompts.

CLI file structure now:

* install.rs — source dispatch + lib wrappers (655 lines)
* install_update.rs — update_capsule + remote version check (343)
* install_prompts.rs — env / elicit prompts (247)
* meta.rs — re-export shim for astrid_capsule_install::meta (10)

New crate layout:

* local.rs — install_from_local_path orchestration
* archive.rs — .capsule tar.gz unpack with traversal/symlink defense
* wasm.rs / wit.rs — BLAKE3 content-addressing
* copy.rs — directory copy excluding *.wasm and wit/
* topics.rs — schema baking
* lifecycle.rs — install/upgrade hook dispatch
* meta.rs — CapsuleMeta + read_meta/write_meta + scan
* paths.rs / manifest_check.rs — small helpers

Tests that exercised the install machinery moved with it
(crates/astrid-capsule-install/tests/install.rs); CLI keeps the
source-resolution tests (GH URL parsing, asset detection,
check_remote_version).

Refs #756
